requirements.txt
certifi==2017.4.17
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.3
django-invitations==1.9.1
django-pydenticon==0.2
django-rest-swagger==2.1.2
djangorestframework==3.6.3
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
mysqlclient==1.3.10
numpy==1.13.1
olefile==0.44
opbeat==3.5.2
openapi-codec==1.3.1
pandas==0.20.3
Pillow==4.2.1
pydenticon==0.3
python-dateutil==2.6.1
python-decouple==3.0
pytz==2017.2
raven==6.1.0
requests==2.18.1
rest-pandas==0.5.0
simplejson==3.11.1
six==1.10.0
uritemplate==3.0.0
urllib3==1.21.1